Move assembly reader and tests to fix dependency
