+++ scripts/update-test-apks.ps1
@@ -4,12 +4,14 @@ Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
 
 $repoRoot = "$PSScriptRoot/.."
-$apkPrefix = "$repoRoot/test/Sentry.Tests/Internals/android"
+$apkDir = Resolve-Path "$repoRoot/test/Sentry.Android.Tests/TestAPKs"
+
+New-Item -ItemType Directory -Force -Path $apkDir | Out-Null
 
 function BuildAndroidSample([bool] $UseAssemblyStore, [bool] $UseAssemblyCompression)
 {
     $sampleDir = "$repoRoot/samples/Sentry.Samples.Android"
-    $outputApk = "$apkPrefix-Store=$UseAssemblyStore-Compressed=$UseAssemblyCompression.apk"
+    $outputApk = "$apkDir/android-Store=$UseAssemblyStore-Compressed=$UseAssemblyCompression.apk"
 
     if ($IfNotExist -and (Test-Path $outputApk))
     {
@@ -21,9 +23,7 @@ function BuildAndroidSample([bool] $UseAssemblyStore, [bool] $UseAssemblyCompres
     try
     {
         # Need to do a clean build otherwise some DLLs would end up being compressed even if it's disabled on this run.
-        git clean -ffxd . | Out-Host
-
-        dotnet build --configuration Release `
+        dotnet build --configuration Release --no-incremental `
             --property:AndroidUseAssemblyStore=$UseAssemblyStore `
             --property:AndroidEnableAssemblyCompression=$UseAssemblyCompression `
         | Out-Host
@@ -37,13 +37,14 @@ function BuildAndroidSample([bool] $UseAssemblyStore, [bool] $UseAssemblyCompres
         Pop-Location
     }
 
-    Move-Item -Verbose "$sampleDir/bin/Release/*/io.sentry.dotnet.samples.android-Signed.apk" $outputApk
+    Move-Item "$sampleDir/bin/Release/*/io.sentry.dotnet.samples.android-Signed.apk" $outputApk
 }
 
 if (!$IfNotExist)
 {
-    Remove-Item -Verbose "$apkPrefix-*.apk"
+    Remove-Item "$apkDir/*.apk"
 }
+
 BuildAndroidSample $true $true
 BuildAndroidSample $true $false
 BuildAndroidSample $false $true
