+++ src/Sentry/Platforms/Android/AssemblyReader/AndroidAssemblyStoreReader.cs
@@ -1,168 +1,7 @@
-// TODO this is actually only needed for Android but we want to test on local devices. Is there another way to do this?
-#if NET6_0_OR_GREATER && !__IOS__
-
 using System.Reflection.PortableExecutable;
 using Sentry.Extensibility;
-using K4os.Compression.LZ4;
 
-namespace Sentry.Internal;
-
-internal interface IAndroidAssemblyReader : IDisposable
-{
-    PEReader? TryReadAssembly(string name);
-}
-
-internal static class AndroidAssemblyReaderFactory
-{
-    public static IAndroidAssemblyReader Open(string apkPath, IList<string> supportedAbis, IDiagnosticLogger? logger)
-    {
-        logger?.LogDebug("Opening APK: {0}", apkPath);
-        var zipArchive = ZipFile.Open(apkPath, ZipArchiveMode.Read);
-
-        if (zipArchive.GetEntry("assemblies/assemblies.manifest") is not null)
-        {
-            logger?.LogDebug("APK uses AssemblyStore");
-            return new AndroidAssemblyStoreReader(zipArchive, supportedAbis, logger);
-        }
-
-        logger?.LogDebug("APK doesn't use AssemblyStore");
-        return new AndroidAssemblyDirectoryReader(zipArchive, supportedAbis, logger);
-    }
-}
-
-internal abstract class AndroidAssemblyReader : IDisposable
-{
-    protected IDiagnosticLogger? Logger { get; }
-    protected ZipArchive ZipArchive { get; }
-    protected IList<string> SupportedAbis { get; }
-
-    protected AndroidAssemblyReader(ZipArchive zip, IList<string> supportedAbis, IDiagnosticLogger? logger)
-    {
-        ZipArchive = zip;
-        Logger = logger;
-        SupportedAbis = supportedAbis;
-    }
-
-    public void Dispose()
-    {
-        ZipArchive.Dispose();
-    }
-
-    protected PEReader CreatePEReader(string assemblyName, MemoryStream inputStream)
-    {
-        var decompressedStream = TryDecompressLZ4(assemblyName, inputStream);
-
-        // Use the decompressed stream, or if null, i.e. it wasn't compressed, use the original.
-        return new PEReader(decompressedStream ?? inputStream);
-    }
-
-    /// <summary>
-    /// The DLL may be LZ4 compressed, see https://github.com/xamarin/xamarin-android/pull/4686
-    /// The format is:
-    ///    [ 4 byte magic header ] (XALZ)
-    ///    [ 4 byte header index ]
-    ///    [ 4 byte uncompressed payload length ]
-    ///    [rest: lz4 compressed payload]
-    /// </summary>
-    /// <seealso href="https://github.com/xamarin/xamarin-android/blob/c92702619f5fabcff0ed88e09160baf9edd70f41/tools/decompress-assemblies/main.cs#L26" />
-    private Stream? TryDecompressLZ4(string assemblyName, MemoryStream inputStream)
-    {
-        const uint compressedDataMagic = 0x5A4C4158; // 'XALZ', little-endian
-        const int payloadOffset = 12;
-        var reader = new BinaryReader(inputStream);
-        if (reader.ReadUInt32() != compressedDataMagic)
-        {
-            // Restore the input stream to the beginning if we're not decompressing.
-            inputStream.Position = 0;
-            return null;
-        }
-        reader.ReadUInt32(); // ignore descriptor index, we don't need it
-        var decompressedLength = reader.ReadInt32();
-        Debug.Assert(inputStream.Position == payloadOffset);
-        var inputLength = (int)(inputStream.Length - payloadOffset);
-
-        Logger?.LogDebug("Decompressing assembly ({0} bytes uncompressed) using LZ4", decompressedLength);
-
-        var outputStream = new MemoryStream(decompressedLength);
-
-        // We're writing to the underlying array manually, so we need to set the length.
-        outputStream.SetLength(decompressedLength);
-        var outputBuffer = outputStream.GetBuffer();
-
-        var inputBuffer = inputStream is MemorySlice slice ? slice.FullBuffer : inputStream.GetBuffer();
-        var offset = inputStream is MemorySlice memorySlice ? memorySlice.Offset + payloadOffset : payloadOffset;
-        var decoded = LZ4Codec.Decode(inputBuffer, offset, inputLength, outputBuffer, 0, decompressedLength);
-        if (decoded != decompressedLength)
-        {
-            throw new Exception($"Failed to decompress LZ4 data of assembly {assemblyName} - decoded {decoded} instead of expected {decompressedLength} bytes");
-        }
-        return outputStream;
-    }
-
-    // Allows consumer to access the underlying buffer even if the MemoryStream is created as a slice over another.
-    // Plain MemoryStream would throw "MemoryStream's internal buffer cannot be accessed."
-    protected class MemorySlice : MemoryStream
-    {
-        public readonly int Offset;
-        public readonly byte[] FullBuffer;
-
-        public MemorySlice(MemoryStream other, int offset, int size) : base(other.GetBuffer(), offset, size, writable: false)
-        {
-            Offset = offset;
-            FullBuffer = other.GetBuffer();
-        }
-    }
-}
-
-// The "Old" app type - where each DLL is placed in the 'assemblies' directory as an individual file.
-internal sealed class AndroidAssemblyDirectoryReader : AndroidAssemblyReader, IAndroidAssemblyReader
-{
-    public AndroidAssemblyDirectoryReader(ZipArchive zip, IList<string> supportedAbis, IDiagnosticLogger? logger)
-        : base(zip, supportedAbis, logger) { }
-
-    public PEReader? TryReadAssembly(string name)
-    {
-        var zipEntry = FindAssembly(name);
-        if (zipEntry is null)
-        {
-            Logger?.LogDebug("Couldn't find assembly {0} in the APK", name);
-            return null;
-        }
-
-        Logger?.LogDebug("Resolved assembly {0} in the APK at {1}", name, zipEntry.FullName);
-
-        // We need a seekable stream for the PEReader (or even to check whether the DLL is compressed), so make a copy.
-        var memStream = new MemoryStream((int)zipEntry.Length);
-        using (var zipStream = zipEntry.Open())
-        {
-            zipStream.CopyTo(memStream);
-            memStream.Position = 0;
-        }
-        return CreatePEReader(name, memStream);
-    }
-
-    private ZipArchiveEntry? FindAssembly(string name)
-    {
-        var zipEntry = ZipArchive.GetEntry($"assemblies/{name}");
-
-        if (zipEntry is null)
-        {
-            foreach (var abi in SupportedAbis)
-            {
-                if (abi.Length > 0)
-                {
-                    zipEntry = ZipArchive.GetEntry($"assemblies/{abi}/{name}");
-                    if (zipEntry is not null)
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
-        return zipEntry;
-    }
-}
+namespace Sentry.Android.AssemblyReader;
 
 // See https://devblogs.microsoft.com/dotnet/performance-improvements-in-dotnet-maui/#single-file-assembly-stores
 internal sealed class AndroidAssemblyStoreReader : AndroidAssemblyReader, IAndroidAssemblyReader
@@ -672,4 +511,3 @@ internal sealed class AndroidAssemblyStoreReader : AndroidAssemblyReader, IAndro
         public MemoryStream? GetAssemblyConfig() => Store.GetAssemblyConfigSlice(this);
     }
 }
-#endif
