+++ src/Sentry/Platforms/Android/AssemblyReader/IAndroidAssemblyReader.cs
@@ -1,0 +1,8 @@
+using System.Reflection.PortableExecutable;
+
+namespace Sentry.Android.AssemblyReader;
+
+internal interface IAndroidAssemblyReader : IDisposable
+{
+    PEReader? TryReadAssembly(string name);
+}
