+++ test/Sentry.Maui.Device.TestApp/Sentry.Maui.Device.TestApp.csproj
@@ -47,14 +47,14 @@
   <!-- Configure XUnit -->
   <ItemGroup>
     <AndroidAsset Include="..\xunit.runner.json" Condition="'$(TargetPlatformIdentifier)' == 'android'" />
-    <Content Include="..\xunit.runner.json" CopyToOutputDirectory="PreserveNewest"
-             Condition="'$(TargetPlatformIdentifier)' == 'ios' Or '$(TargetPlatformIdentifier)' == 'maccatalyst'" />
+    <Content Include="..\xunit.runner.json" CopyToOutputDirectory="PreserveNewest" Condition="'$(TargetPlatformIdentifier)' == 'ios' Or '$(TargetPlatformIdentifier)' == 'maccatalyst'" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\MauiTestUtils\DeviceTests\TestUtils.DeviceTests.csproj" />
     <ProjectReference Include="..\MauiTestUtils\DeviceTests.Runners\TestUtils.DeviceTests.Runners.csproj" />
     <ProjectReference Include="..\MauiTestUtils\DeviceTests.Runners.SourceGen\TestUtils.DeviceTests.Runners.SourceGen.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Sentry.Android.Tests\Sentry.Android.Tests.csproj" Condition="'$(TargetPlatformIdentifier)' == 'android'" />
     <ProjectReference Include="..\Sentry.Tests\Sentry.Tests.csproj" />
     <ProjectReference Include="..\Sentry.Extensions.Logging.Tests\Sentry.Extensions.Logging.Tests.csproj" />
     <ProjectReference Include="..\Sentry.Maui.Tests\Sentry.Maui.Tests.csproj" />
