+++ test/Sentry.Maui.Device.TestApp/Startup.cs
@@ -14,7 +14,11 @@ public static class MauiProgram
                 {
                     typeof(Sentry.Tests.SentrySdkTests).Assembly,
                     typeof(Sentry.Extensions.Logging.Tests.LogLevelExtensionsTests).Assembly,
-                    typeof(Sentry.Maui.Tests.SentryMauiOptionsTests).Assembly
+                    typeof(Sentry.Maui.Tests.SentryMauiOptionsTests).Assembly,
+
+#if ANDROID
+                    typeof(Sentry.Android.Tests.AndroidAssemblyReaderTests).Assembly,
+#endif
                 }
             })
             .UseHeadlessRunner(new HeadlessRunnerOptions
